more accurate type definition
