+++ packages/chain-adapters/package.json
@@ -31,7 +31,7 @@
     "bech32": "^2.0.0",
     "bignumber.js": "^9.0.2",
     "bitcoinjs-lib": "^5.2.0",
-    "coinselect": "^3.1.12",
+    "coinselect": "^3.1.13",
     "ethers": "^5.4.7",
     "multicoin-address-validator": "^0.5.2",
     "web3-utils": "^1.7.3"
